% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_process_subfuncs.R
\name{compute_evp}
\alias{compute_evp}
\title{EVP Base Function}
\usage{
compute_evp(
  cohort,
  grouped_list,
  time = FALSE,
  evp_variable_file = read_codeset("evp_variables", "cccc")
)
}
\arguments{
\item{cohort}{table of cohort members with at least \code{site}, \code{person_id}, \code{start_date}, and \code{end_date}}

\item{grouped_list}{list of columns that should be used to group the table}

\item{time}{logical to determine whether the function is being run as part of \code{compute_fot} or not}

\item{evp_variable_file}{CSV file with information about each of the variables that should be
examined in the function. contains the following columns:

\if{html}{\out{<div class="sourceCode">}}\preformatted{                    `concept_group`, `default_tbl`, `field_name`, `date_field`, `codeset_name`,
                    `filter_logic`
}\if{html}{\out{</div>}}}
}
\value{
dataframe with patient/row counts and proportions that are computed per group defined in
grouped_list, and if time = TRUE, for each time period defined in compute_fot
}
\description{
EVP Base Function
}
