% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_process_subfuncs.R
\name{compute_evp_ssanom}
\alias{compute_evp_ssanom}
\title{Single site anomaly no time processing for EVP}
\usage{
compute_evp_ssanom(
  cohort,
  grouped_list,
  evp_variable_file = read_codeset("evp_variables", "cccc")
)
}
\arguments{
\item{cohort}{table of cohort members with at least \code{site}, \code{person_id}, \code{start_date}, and \code{end_date}}

\item{grouped_list}{list of columns that should be used to group the table}

\item{evp_variable_file}{CSV file with information about each of the variables that should be
examined in the function. contains the following columns:

\if{html}{\out{<div class="sourceCode">}}\preformatted{                    `concept_group`, `default_tbl`, `field_name`, `date_field`, `codeset_name`,
                    `filter_logic`
}\if{html}{\out{</div>}}}
}
\value{
one dataframe with the jaccard similarity index for each concept group provided
in the concept file
}
\description{
Single site anomaly no time processing for EVP
}
