% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_process_subfuncs_omop.R
\name{compute_evp_ssanom_omop}
\alias{compute_evp_ssanom_omop}
\title{Single site anomaly no time processing for EVP}
\usage{
compute_evp_ssanom_omop(
  cohort,
  grouped_list,
  evp_variable_file = expectedvariablespresent::evp_variable_file_omop
)
}
\arguments{
\item{cohort}{table of cohort members with at least \code{site}, \code{person_id}, \code{start_date}, and \code{end_date}}

\item{grouped_list}{list of columns that should be used to group the table}

\item{evp_variable_file}{CSV file with information about each of the variables that should be
examined in the function. contains the following columns:
\itemize{
\item \code{variable} a label for the variable captured by the associated codeset
\item \code{default_tbl} CDM table where data related to the codeset is found
\item \code{concept_field} concept_id field with codes from the associated codeset
\item \code{date_field} a date field in the \code{default_tbl} that should be used for over time analyses
\item \code{codeset_name} the name of the codeset file; DO NOT include the file extension
\item \code{filter_logic} a string indicating filter logic that should be applied to achieve the desired variable; optional
}}
}
\value{
one dataframe with the jaccard similarity index for each concept group provided
in the concept file
}
\description{
Single site anomaly no time processing for EVP
}
\keyword{internal}
