% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_output_subfuncs.R
\name{evp_ms_anom_cs}
\alias{evp_ms_anom_cs}
\title{\itemize{
\item Multi Site, Anomaly, Cross-Sectional*
}}
\usage{
evp_ms_anom_cs(process_output, output_level, text_wrapping_char = 60)
}
\arguments{
\item{process_output}{the output provided by the \code{evp_process} function}

\item{output_level}{the level of output to be displayed: \code{patient} or \code{row}}

\item{text_wrapping_char}{integer value indicating the point at which axis
label text should begin to wrap}
}
\value{
a dot plot where the shape of the dot represents whether the point is
anomalous, the color of the dot represents the proportion of rows/patients
for a given variable, and the size of the dot represents the mean proportion
across all sites

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if there were no groups eligible for analysis, a heat map showing the proportion
    and a dot plot showing each site's average standard deviation away from the mean
    proportion is returned instead
}\if{html}{\out{</div>}}
}
\description{
\itemize{
\item Multi Site, Anomaly, Cross-Sectional*
}
}
