% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_output_subfuncs.R
\name{evp_ms_anom_la}
\alias{evp_ms_anom_la}
\title{\strong{Multi-Site, Anomaly, Longitudinal}}
\usage{
evp_ms_anom_la(
  process_output,
  output_level,
  filter_variable,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{output from \code{evp_process}}

\item{output_level}{the level of output that should be shown (\code{person} or \code{row})}

\item{filter_variable}{the variable that should be used to generate output}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
three graphs:
\enumerate{
\item line graph that shows the smoothed proportion of a
variable across time computation with the Euclidean distance associated with each line
\item line graph that shows the raw proportion of a
variable across time computation with the Euclidean distance associated with each line
\item a bar graph with the Euclidean distance value for each site, with the average
proportion as the fill
}

THIS GRAPH SHOWS ONLY ONE VARIABLE AT A TIME!
}
\description{
\strong{Multi-Site, Anomaly, Longitudinal}
}
