% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_output_subfuncs.R
\name{evp_ms_exp_cs}
\alias{evp_ms_exp_cs}
\title{\itemize{
\item Multi Site, Exploratory, Cross-Sectional*
}}
\usage{
evp_ms_exp_cs(process_output, output_level, facet = NULL)
}
\arguments{
\item{process_output}{the output provided by the \code{evp_process} function}

\item{output_level}{the level of output to be displayed: \code{patient} or \code{row}}

\item{facet}{columns the user would like to facet by}
}
\value{
a heat map displaying the proportion of patients/rows that meet criteria for each
of the variables found in process_output at each of site
}
\description{
\itemize{
\item Multi Site, Exploratory, Cross-Sectional*
}
}
