% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_output_subfuncs.R
\name{evp_ms_exp_la}
\alias{evp_ms_exp_la}
\title{\itemize{
\item Multi Site, Exploratory, Longitudinal*
}}
\usage{
evp_ms_exp_la(
  process_output,
  output_level,
  filter_variable,
  large_n = FALSE,
  large_n_sites = NULL,
  facet = NULL
)
}
\arguments{
\item{process_output}{the output provided by the \code{evp_process} function}

\item{output_level}{the level of output to be displayed: \code{patient} or \code{row}}

\item{filter_variable}{the single variable the output should display}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}

\item{facet}{columns the user would like to facet by}
}
\value{
a line graph displaying the proportion of patients/rows for each variable
& site over the user-specified time period.
}
\description{
\itemize{
\item Multi Site, Exploratory, Longitudinal*
}
}
\keyword{internal}
