% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_process.R
\name{evp_process}
\alias{evp_process}
\title{Expected Variables Present}
\usage{
evp_process(
  cohort,
  omop_or_pcornet,
  evp_variable_file,
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  output_level = "patient",
  age_groups = NULL,
  p_value = 0.9,
  time = FALSE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year"
)
}
\arguments{
\item{cohort}{\emph{tabular input} || \strong{required}

The cohort to be used for data quality testing. This table should contain,
at minimum:
\itemize{
\item \code{site} | \emph{character} | the name(s) of institutions included in your cohort
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character} | the patient identifier
\item \code{start_date} | \emph{date} | the start of the cohort period
\item \code{end_date} | \emph{date} | the end of the cohort period
}

Note that the start and end dates included in this table will be used to
limit the search window for the analyses in this module.}

\item{omop_or_pcornet}{\emph{string} || \strong{required}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data
\itemize{
\item \code{omop}: run the \code{\link[=evp_process_omop]{evp_process_omop()}} function against an OMOP CDM instance
\item \code{pcornet}: run the \code{\link[=evp_process_pcornet]{evp_process_pcornet()}} function against a PCORnet CDM instance
}}

\item{evp_variable_file}{\emph{tabular input} || \strong{required}

A table with information about each of the variables that should be examined
in the analysis. This table should contain the following columns:
\itemize{
\item \code{variable} | \emph{character} | a string label for the variable captured by the associated codeset
\item \code{domain_tbl} | \emph{character} | the CDM table where the variable is found
\item \code{concept_field} | \emph{character} | the string name of the field in the domain table where the concepts are located
\item \code{date_field} | \emph{character} | the name of the field in the domain table with the date that should be used for temporal filtering
\item \code{vocabulary_field} | \emph{character} | for PCORnet applications, the name of the field in the domain table with a vocabulary identifier to differentiate concepts from one another (ex: dx_type); can be set to NA for OMOP applications
\item \code{codeset_name} | \emph{character} | the name of the codeset that defines the variable of interest
\item \code{filter_logic} | \emph{character} | logic to be applied to the domain_tbl in order to achieve the definition of interest; should be written as if you were applying it in a dplyr::filter command in R
}

To see an example of the structure of this file, please see \code{?expectedvariablespresent::evp_variable_file_omop} or
\code{?expectedvariablespresent::evp_variable_file_pcornet}}

\item{multi_or_single_site}{\emph{string} || defaults to \code{single}

A string, either \code{single} or \code{multi}, indicating whether a single-site or
multi-site analysis should be executed}

\item{anomaly_or_exploratory}{\emph{string} || defaults to \code{exploratory}

A string, either \code{anomaly} or \code{exploratory}, indicating what type of results
should be produced.

Exploratory analyses give a high level summary of the data to examine the
fact representation within the cohort. Anomaly detection analyses are
specialized to identify outliers within the cohort.}

\item{output_level}{\emph{string} || defaults to \code{patient}

A string indicating the analysis level to use as the basis for the
Multi Site, Anomaly Detection computations

Acceptable values are either \code{patient} or \code{row}}

\item{age_groups}{\emph{tabular input} || defaults to \code{NULL}

If you would like to stratify the results by age group, create a table or
CSV file with the following columns and use it as input to this parameter:
\itemize{
\item \code{min_age} | \emph{integer} | the minimum age for the group (i.e. 10)
\item \code{max_age} | \emph{integer} | the maximum age for the group (i.e. 20)
\item \code{group} | \emph{character} | a string label for the group (i.e. 10-20, Young Adult, etc.)
}

If you would \emph{not} like to stratify by age group, leave as \code{NULL}}

\item{p_value}{\emph{numeric} || defaults to \code{0.9}

The p value to be used as a threshold in the Multi-Site,
Anomaly Detection, Cross-Sectional analysis}

\item{time}{\emph{boolean} || defaults to \code{FALSE}

A boolean to indicate whether to execute a longitudinal analysis}

\item{time_span}{\emph{vector - length 2} || defaults to \code{c('2012-01-01', '2020-01-01')}

A vector indicating the lower and upper bounds of the time series for longitudinal analyses}

\item{time_period}{\emph{string} || defaults to \code{year}

A string indicating the distance between dates within the specified time_span.
Defaults to \code{year}, but other time periods such as \code{month} or \code{week} are
also acceptable}
}
\value{
This function will return a dataframe summarizing the
distribution of each user-defined variable. For a
more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
This is a completeness module that will assess the presence of expected study variables and compute the distribution of
these variables in the dataset. The user will provide the variables (\code{evp_variable_file}) of interest, including the name of
the concept sets with the concepts used to define the variable. Sample versions of these inputs, both for OMOP and
PCORnet, are included as data in the package and are accessible with \verb{expectedvariablespresent::}.
Results can optionally be stratified by site, age group, and/or time. This function is compatible with
both the OMOP and the PCORnet CDMs based on the user's selection.
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'expectedvariablespresent'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'evp_process_test',
                      working_directory = my_directory,
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = my_file_folder,
                      cdm_schema = NA)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000), # RSQLite does not store date objects,
                                      # hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Execute `evp_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
evp_process_example <- evp_process(cohort = cohort,
                                   multi_or_single_site = 'single',
                                   anomaly_or_exploratory = 'exploratory',
                                   time = FALSE,
                                   omop_or_pcornet = 'omop',
                                   evp_variable_file = evp_variable_file_omop) \%>\%
  suppressMessages()

evp_process_example

#' Execute `evp_output` function
evp_output_example <- evp_output(process_output = evp_process_example,
                                 output_level = 'patient')

evp_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(evp_output_example)
}
