% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_process_pcornet.R
\name{evp_process_pcornet}
\alias{evp_process_pcornet}
\title{EVP Process function}
\usage{
evp_process_pcornet(
  cohort,
  evp_variable_file = read_codeset("evp_variables", "cccc"),
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  output_level = "row",
  age_groups = NULL,
  p_value = 0.9,
  time = FALSE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year"
)
}
\arguments{
\item{cohort}{A dataframe with the cohort of patients for your study. Should include the columns:
- @person_id
- @start_date
- @end_date
- @site}

\item{evp_variable_file}{CSV file with information about each of the variables that should be
examined in the function. contains the following columns:
- @variable a label for the variable captured by the associated codeset
- @default_tbl CDM table where data related to the codeset is found
- @field_name concept_id field with codes from the associated codeset
- @date_field a date field in the \code{default_tbl} that should be used for
over time analyses
- @codeset_name the name of the codeset file; DO NOT include the file extension}

\item{multi_or_single_site}{Option to run the function on a single vs multiple sites
- @single - run the function for a single site
- @multi - run the function for multiple sites}

\item{anomaly_or_exploratory}{Option to conduct an exploratory or anomaly detection analysis. Exploratory analyses give a high
level summary of the data to examine the fact representation within the cohort. Anomaly detection
analyses are specialized to identify outliers within the cohort.}

\item{output_level}{the level of output to use for an AUC computation, exclusive to \code{ms_anom_at}; either \code{patient} or \code{row} --
defaults to \code{row}}

\item{age_groups}{If you would like to stratify the results by age group, fill out the provided \code{age_group_definitions.csv} file
with the following information:
- @min_age: the minimum age for the group (i.e. 10)
- @max_age: the maximum age for the group (i.e. 20)
- @group: a string label for the group (i.e. 10-20, Young Adult, etc.)

\if{html}{\out{<div class="sourceCode">}}\preformatted{                Then supply this csv file as the age_groups argument (i.e. read.csv('path/to/age_group_definitions.csv'))

                If you would not like to stratify by age group, leave the argument as NULL
}\if{html}{\out{</div>}}}

\item{p_value}{the p value to be used as a threshold in the multi-site anomaly detection analysis}

\item{time}{a logical that tells the function whether you would like to look at the output over time}

\item{time_span}{when time = TRUE, this argument defines the start and end dates for the time period of interest. should be
formatted as c(start date, end date) in yyyy-mm-dd date format}

\item{time_period}{when time = TRUE, this argument defines the distance between dates within the specified time period. defaults
to \code{year}, but other time periods such as \code{month} or \code{week} are also acceptable}
}
\value{
a dataframe with patient/row counts & proportions for each concept set listed in \code{evp_concept_file}.
this output should then be used in the \code{evp_output} function to generate an appropriate
visualization
}
\description{
EVP Process function
}
