% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_output_subfuncs.R
\name{evp_ss_anom_cs}
\alias{evp_ss_anom_cs}
\title{\itemize{
\item Single Site, Anomaly, Cross-Sectional*
}}
\usage{
evp_ss_anom_cs(process_output, facet = NULL)
}
\arguments{
\item{process_output}{the output provided by the \code{evp_process} function}

\item{facet}{columns the user would like to facet by}
}
\value{
a heat map displaying the Jaccard similarity index between each of the variables
any variables without a relationship and any self-to-self relationships are dropped
}
\description{
\itemize{
\item Single Site, Anomaly, Cross-Sectional*
}
}
