% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_output_subfuncs.R
\name{evp_ss_anom_la}
\alias{evp_ss_anom_la}
\title{\itemize{
\item Single Site, Anomaly, Longitudinal*
}}
\usage{
evp_ss_anom_la(process_output, output_level, filter_variable, facet = NULL)
}
\arguments{
\item{process_output}{the output provided by the \code{evp_process} function}

\item{output_level}{the level of output to be displayed: \code{patient} or \code{row}}

\item{filter_variable}{the single variable the output should display}

\item{facet}{columns the user would like to facet by}
}
\value{
if analysis was executed by year or greater, a P Prime control chart
is returned with outliers marked with orange dots

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if analysis was executed by month or smaller, an STL regression is
    conducted and outliers are marked with red dots. the graphs representing
    the data removed in the regression are also returned
}\if{html}{\out{</div>}}
}
\description{
\itemize{
\item Single Site, Anomaly, Longitudinal*
}
}
