% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_output_subfuncs.R
\name{evp_ss_anom_nt}
\alias{evp_ss_anom_nt}
\title{\itemize{
\item Single Site, Anomaly, No Time *
}}
\usage{
evp_ss_anom_nt(process_output, facet = NULL)
}
\arguments{
\item{process_output}{the output provided by the \code{evp_process} function}

\item{facet}{columns the user would like to facet by}
}
\value{
a heat map displaying the Jaccard similarity index between each of the variables
any variables without a relationship and any self-to-self relationships are dropped

.        if the user hovers over the heatmap, the co-occurrence count, jaccard score for the pair,
mean jaccard score for the variable, and variables will show
}
\description{
\itemize{
\item Single Site, Anomaly, No Time *
}
}
