% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_output_subfuncs.R
\name{evp_ss_exp_la}
\alias{evp_ss_exp_la}
\title{\itemize{
\item Single Site, Exploratory, Longitudinal*
}}
\usage{
evp_ss_exp_la(process_output, output_level, facet = NULL)
}
\arguments{
\item{process_output}{the output provided by the \code{evp_process} function}

\item{output_level}{the level of output to be displayed: \code{patient} or \code{row}}

\item{facet}{columns the user would like to facet by}
}
\value{
a line graph displaying the proportion of patients/rows for each variable
over the user-specified time period
}
\description{
\itemize{
\item Single Site, Exploratory, Longitudinal*
}
}
\keyword{internal}
