% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evp_variable_file_pcornet}
\alias{evp_variable_file_pcornet}
\title{EVP Sample Variable File -- PCORnet}
\format{
\subsection{\code{evp_variable_file_pcornet}}{

A data frame with 2 rows and 6 columns
\describe{
\item{variable}{A string label for the variable of interest}
\item{domain_tbl}{The name of the CDM table where the variable can be found}
\item{concept_field}{The field in the default_tbl that should be used to join to the codeset}
\item{date_field}{The date field in the default_tbl that should be used to filter the dataset to the cohort period and for longitudinal analyses}
\item{vocabulary_field}{field in the \code{default_tbl} that defines the vocabulary type of the concept (i.e. dx_type)}
\item{codeset_name}{The name of the codeset as found in the specs directory; file extension should not be included}
\item{filter_logic}{optional; a string to be parsed as logic to filter the default_tbl and better identify the variable of interest}
}
}
}
\usage{
evp_variable_file_pcornet
}
\description{
A sample version of the file structure expected for the evp_variable_file
parameter in the \code{evp_process} function. The user should recreate this file
structure and include their own variables.
}
\details{
Please note that the codesets should be stored in the \code{file_subdirectory} established
when \code{ssdqa.gen::initialize_dq_session} is executed.

Examples of appropriately structured codeset files are attached to the ssdqa.gen
package and can be accessed with \verb{ssdqa.gen::}. Please note that if the vocabulary_field
is not NULL (i.e. for diagnosis or procedures concepts), the vocabulary_id field in the codeset
must be included.
}
\keyword{datasets}
