% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedvariablespresent-package.R
\docType{package}
\name{expectedvariablespresent-package}
\alias{expectedvariablespresent}
\alias{expectedvariablespresent-package}
\title{expectedvariablespresent: Expected Variables Present}
\description{
This package represents one of several modules in the broader SSDQA ecosystem. The Expected Variables Present module measures the presence of user-defined study variables within the source data.
}
\author{
\strong{Maintainer}: Kaleigh Wieand \email{wieandk@chop.edu}

}
\keyword{internal}
