% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_output.R
\name{evp_output}
\alias{evp_output}
\title{Expected Variables Present -- Output Generation}
\usage{
evp_output(
  process_output,
  output_function,
  output_level,
  filter_variable = NULL
)
}
\arguments{
\item{process_output}{the output of the \code{evp_process} function}

\item{output_function}{the name of the output function that should be used provided in the \code{parameter_summary} csv
file that is output to the provided results folder after running the \code{evp_process} function}

\item{output_level}{the type of counts the output should summarise -- either \code{patient} or \code{row}}

\item{filter_variable}{for \code{evp_ms_anom_la}, \code{evp_ms_exp_la}, \code{evp_ss_anom_la}, the single variable that
should be displayed in the output; can be any of the variables listed in the \code{evp_process} output}
}
\value{
a graph to visualize the results from \code{evp_process} based on the parameters provided; see documentation
for individual subfunctions for details on specific output
}
\description{
Using the tabular output generated by \code{evp_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_squba()}
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'expectedvariablespresent'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'evp_process_test',
                      working_directory = getwd(),
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = system.file('extdata',
                                        package = 'expectedvariablespresent'),
                      cdm_schema = NA)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000), # RSQLite does not store date objects,
                                      # hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Execute `evp_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
evp_process_example <- evp_process(cohort = cohort,
                                   multi_or_single_site = 'single',
                                   anomaly_or_exploratory = 'exploratory',
                                   time = FALSE,
                                   omop_or_pcornet = 'omop',
                                   evp_variable_file = evp_variable_file_omop)

evp_process_example

#' Execute `evp_output` function
evp_output_example <- evp_output(process_output = evp_process_example,
                                 output_function = 'evp_ss_exp_cs',
                                 output_level = 'patient')

evp_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(evp_output_example)
}
