% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_output.R
\name{evp_output}
\alias{evp_output}
\title{Expected Variables Present -- Output Generation}
\usage{
evp_output(
  process_output,
  output_level = "patient",
  filter_variable = NULL,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{\emph{tabular input} || \strong{required}

The tabular output produced by \code{evp_process}}

\item{output_level}{\emph{string} || defaults to \code{patient}

A string indicating the analysis level that should be reflected in the plot. All checks
utilize this parameter EXCEPT \verb{Single Site, Anomaly Detection, Cross-Sectional}, which
executes a Jaccard index using only patient counts}

\item{filter_variable}{\emph{string or vector} || defaults to \code{NULL}

A string or vector with variable names that should be the focus of the analysis.
Only a single value is accepted for the following checks:
\itemize{
\item \verb{Multi Site, Exploratory, Longitudinal} (non-year time period)
\item \verb{Multi Site, Anomaly Detection, Longitudinal}
\item \verb{Single Site, Anomaly Detection, Longitudinal}
}

Multiple values (up to 3) are accepted for:
\itemize{
\item \verb{Multi Site, Exploratory, Longitudinal} (year time period)
}}

\item{large_n}{\emph{boolean} || defaults to \code{FALSE}

For Multi-Site analyses, a boolean indicating whether the large N
visualization, intended for a high volume of sites, should be used. This
visualization will produce high level summaries across all sites, with an
option to add specific site comparators via the \code{large_n_sites} parameter.}

\item{large_n_sites}{\emph{vector} || defaults to \code{NULL}

When \code{large_n = TRUE}, a vector of site names that can add site-level information
to the plot for comparison across the high level summary information.}
}
\value{
This function will produce a graph to visualize the results
from \code{evp_process} based on the parameters provided. The default
output is typically a static ggplot or gt object, but interactive
elements can be activated by passing the plot through \code{make_interactive_squba}.
For a more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
Using the tabular output generated by \code{evp_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_squba()}
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'expectedvariablespresent'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'evp_process_test',
                      working_directory = my_directory,
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = my_file_folder,
                      cdm_schema = NA)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000), # RSQLite does not store date objects,
                                      # hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Execute `evp_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
evp_process_example <- evp_process(cohort = cohort,
                                   multi_or_single_site = 'single',
                                   anomaly_or_exploratory = 'exploratory',
                                   time = FALSE,
                                   omop_or_pcornet = 'omop',
                                   evp_variable_file = evp_variable_file_omop) \%>\%
  suppressMessages()

evp_process_example

#' Execute `evp_output` function
evp_output_example <- evp_output(process_output = evp_process_example,
                                 output_level = 'patient')

evp_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(evp_output_example)
}
