% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_output.R
\name{evp_output}
\alias{evp_output}
\title{EVP output function}
\usage{
evp_output(
  process_output,
  output_function,
  output_level,
  filter_variable,
  facet = NULL
)
}
\arguments{
\item{process_output}{the output of the \code{evp_process} function}

\item{output_function}{the name of the output function that should be used provided in the \code{parameter_summary} csv
file that is output to the provided results folder after running the \code{evp_process} function}

\item{output_level}{the type of counts the output should summarise -- either \code{patient} or \code{row}}

\item{filter_variable}{for @ms_anom_at, @ms_exp_at, and @ss_anom_at, the single variable that should be displayed in the output;
can be any of the variables listed in the \code{evp_process} output}

\item{facet}{the variables by which you would like to facet the graph. available and/or recommended options for
faceting variables are provided in the \code{parameter_summary} csv file}
}
\value{
a graph to visualize the results from \code{evp_process} based on the parameters provided; see documentation
for individual subfunctions for details on specific output
}
\description{
EVP output function
}
